requirements.txt
uwsgi==2.0.15
django==1.11.8
django-registration==2.3
django-registration-redux==1.9
django-settings-export==1.2.1
django-ckeditor==5.3.1
django-forms-bootstrap==3.1.0
django-rosetta==0.7.13
django-template-check==0.3.0
django-dia==0.3.4
graphviz==0.8.1
